Recognise upper case username fields
This change adds so that the upper case versions of the username field are
recognised.

The example on the official password-store web page includes "Username", i.e.
upper case U, which is why I've used it across all my files. It seems like a
simple enough change. An alternative would be to programmatically generate the
upper case versions but this is by far easier.